+++ views/_interface.hbs
@@ -1,0 +1,141 @@
+<div class="container col-sm-12" ng-controller="mainCtrl">
+  <span ng-init="api.pid='{{pid}}';api.mid='{{mid}}'"><a href="/module/index.html?projectName={{projectName}}#{{pid}}">{{projectName}}</a>&nbsp;>&nbsp;{{moduleName}}</span>
+  <h3>
+		<small class="text-info" ng-bind="'版本:'" ng-init="api.version=1;api.author='{{userName}}';">
+		</small>
+		<small ng-bind="api.version">
+			<!-- <select style="width:50px;" ng-model="api.version" ng-change="change();" ng-init="versions=[1]">
+				<option ng-repeat="v in versions" value="//v//" ng-bind="//v//"></option>
+			</select> -->
+		</small>
+		&nbsp;&nbsp;
+		&nbsp;&nbsp;
+		<small class="text-info" ng-bind="'更新日期:'+api.updateDate" ng-init="api.updateDate='{{updateDate}}'"></small>
+	</h3>
+    <form name="form">
+      <div class="form-group">
+        <input type="text" name="name" class="form-control" ng-model="api.name" placeholder="接口名称" validator="required">
+      </div>
+      <div class="form-group">
+        <input type="text" name="url" class="form-control" ng-model="api.url" placeholder='请求URL(必须以"/"开头)' validator="required">
+      </div>
+      <div class="form-group">
+        <div class="col-sm-2" style="line-height:34px;">
+          <label for="" class="control-label">请求方式</label>
+        </div>
+        <div class="col-sm-3">
+          <select name="method" class="form-control" ng-model="api.method" ng-init="api.method='get'">
+            <option value="get">get</option>
+            <option value="post">post</option>
+            <option value="put">put</option>
+            <option value="delete">delete</option>
+          </select>
+        </div>
+        <div ng-if="api.referenceId" class="col-sm-2" style="line-height:34px;">
+		  		<label class="control-label" ng-init="api.authorize=false;">前置接口</label>
+				</div>
+				<div ng-if="api.referenceId" class="col-sm-3" ng-bind="api.referenceName" style="line-height:34px;">
+				</div>
+        <div class="clearfix"></div>
+      </div>
+      <div class="form-group">
+        <table class="table table-bordered table-hover">
+          <tr>
+            <td colspan="5" class="text-center" ng-init="api.inParams=[];">请求参数说明</td>
+          </tr>
+          <tr>
+            <th class="col-sm-2 text-center">参数名</th>
+            <th class="col-sm-2 text-center">是否必传</th>
+            <th class="col-sm-2 text-center">类型</th>
+            <th class="text-center">说明</th>
+            <th class="col-sm-1 text-center"></th>
+          </tr>
+          <tr ng-repeat="p in api.inParams">
+            <td>
+              <input type="text" class="form-control" ng-model="p.name" title="//p.name//" ng-keydown="keydown($event,api.inParams);">
+            </td>
+            <td>
+              <select name="" id="" class="form-control" ng-model="p.isNeed">
+                <option value="true">是</option>
+                <option value="false">否</option>
+              </select>
+            </td>
+            <td>
+              <select name="" class="form-control" ng-model="p.type">
+                <option value="string">string</option>
+                <option value="number">number</option>
+                <option value="double">double</option>
+                <option value="boolean">boolean</option>
+                <option value="object">object</option>
+                <option value="array">array</option>
+                <option value="stream">stream</option>
+              </select>
+            </td>
+            <td>
+              <input type="text" class="form-control" ng-model="p.desc" title="//p.desc//" ng-keydown="keydown($event,api.inParams);">
+            </td>
+            <td class="text-center"><span class="label label-danger btn" ng-dblclick="delParam(p.seq, api.inParams)">-</span></td>
+          </tr>
+          <tr>
+            <td colspan="5" class="text-center">
+              <span class="label label-info btn" ng-click="addParam(api.inParams)">&nbsp;&nbsp;+&nbsp;&nbsp;</span>
+            </td>
+          </tr>
+        </table>
+      </div>
+      <div class="form-group">
+        <textarea name="inObject" class="form-control" rows="10" ng-change="print(api.inObject)" ng-model="api.inObject" placeholder="请求示例（JSON格式）" validator="json"></textarea>
+      </div>
+      <div class="form-group">
+        <table class="table table-bordered table-hover">
+          <tr>
+            <td colspan="5" class="text-center" ng-init="api.outParams=[];">返回参数说明</td>
+          </tr>
+          <tr>
+            <th class="col-sm-2 text-center">参数名</th>
+            <th class="col-sm-2 text-center">类型</th>
+            <th class="text-center">说明</th>
+            <th class="col-sm-1 text-center"> </th>
+          </tr>
+          <tr ng-repeat="p in api.outParams">
+            <td>
+              <input type="text" class="form-control" ng-model="p.name" title="//p.name//" ng-keydown="keydown($event, api.outParams)">
+            </td>
+            <td>
+              <select name="" id="" class="form-control" ng-model="p.type">
+                <option value="string">String</option>
+                <option value="number">Number</option>
+                <option value="double">Double</option>
+                <option value="boolean">Boolean</option>
+                <option value="object">object</option>
+                <option value="array">array</option>
+              </select>
+            </td>
+            <td>
+              <input type="text" class="form-control" ng-model="p.desc" title="//p.desc//" ng-keydown="keydown($event, api.outParams)">
+            </td>
+            <td class="text-center"><span class="label label-danger btn" ng-dblclick="delParam(p.seq, api.outParams)">-</span></td>
+          </tr>
+          <tr>
+            <td colspan="5" class="text-center">
+              <span class="label label-info btn" ng-click="addParam(api.outParams)">&nbsp;&nbsp;+&nbsp;&nbsp;</span>
+            </td>
+          </tr>
+        </table>
+      </div>
+      <div class="form-group">
+        <textarea name="outObject" class="form-control" ng-model="api.outObject" placeholder="返回示例（JSON或字符串）" rows="10" validator="required,jsonString"></textarea>
+      </div>
+      <div class="form-group">
+        <textarea name="remark" class="form-control" placeholder="备注" ng-model="api.remark"></textarea>
+      </div>
+      <div ng-if="message" class="alert alert-//message.level//" role="alert">
+        <span class="glyphicon glyphicon-exclamation-sign" aria-hidden="true"></span>
+        <span class="sr-only">Error:</span> //message.text//
+      </div>
+      <div class="form-group text-right">
+        <button type="button" class="pull-left btn btn-primary btn-large" ng-click="sendRequest();">发送请求</button>
+        <button validation-submit="form" ng-click="submit(form);" class="btn btn-large btn-warning">保存</button>&emsp;&emsp;
+      </div>
+    </form>
+</div>
