+++ views/layout.hbs
@@ -6,7 +6,7 @@
     <link rel="stylesheet" href="/stylesheets/style.css" />
   </head>
   {{#if static}}
-  <body>
+  <body ng-cloak>
   {{else}}
   <body ng-cloak>
   {{/if}}
