+++ views/module.hbs
@@ -14,7 +14,7 @@
         <span ng-bind="m.name" ng-hide="m.text" placeholder="模块名称"></span>
         {{#if editable}}
           <input id="_moduleName" type="text" ng-model="m.text" ng-if="m.text" ng-keydown="keydown($event, m);" ng-blur="cancel(m)">
-          <a href="javascript:void(0);" title="双击删除" ng-hide="m.text" ng-dblclick="remove(m._id);" class="pull-right text-danger glyphicon glyphicon-minus">删除&nbsp;</a>
+          <a href="javascript:void(0);" title="双击删除" ng-hide="m.text" ng-dblclick="remove(m._id);" class="pull-right text-danger glyphicon glyphicon-minus">双击删除&nbsp;</a>
           <a href="javascript:void(0);" ng-hide="m.text" ng-click="edit(m)" class="pull-right text-warning glyphicon glyphicon-edit">编辑&nbsp;</a>
         {{/if}}
       </div>
@@ -34,7 +34,7 @@
           </tr>
           <tr ng-repeat="i in m.interfaces">
             <td>
-              <a target="_blank" href="/schema/index.html?pid=//m.pid//&mid=//m._id//&projectName={{projectName}}&moduleName=//m.name//#//i._id//" ng-bind="i.name+'('+i.version+')'"></a>
+              <a target="_blank" href="/interface/index.html?pid=//m.pid//&mid=//m._id//&projectName=//projectName//&moduleName=//m.name//#//i._id//" ng-bind="i.name+'('+i.version+')'"></a>
             </td>
             <td ng-bind="i.url"></td>
             <td ng-bind="i.method"></td>
@@ -45,20 +45,20 @@
             <td><a ng-if="i.outObject" href="" ng-click="showParam(i.name, i.outObject)" data-toggle="modal" data-target="#myModal">返回值</a></td>
             {{#if editable}}
               <td style="text-align:right;">
-                <a target="_blank" href="/schema/index.html?pid=//m.pid//&mid=//m._id//&projectName={{projectName}}&moduleName=//m.name//#//i._id//" class="text-warning"><span class="glyphicon glyphicon-edit"></span>编辑</a>
-                <a href="" title="双击删除" ng-dblclick="deleteInterface(i._id, m.interfaces)" class="text-danger"><span class="glyphicon glyphicon-minus"></span>删除</a>
+                <a target="_blank" href="/interface/index.html?pid=//m.pid//&mid=//m._id//&projectName=//projectName//&moduleName=//m.name//#//i._id//" class="text-warning"><span class="glyphicon glyphicon-edit"></span>编辑</a>
+                <a href="" title="双击删除" ng-dblclick="deleteInterface(i._id, m.interfaces)" class="text-danger"><span class="glyphicon glyphicon-minus"></span>双击删除</a>
               </td>
             {{/if}}
             <td >
               <span data-toggle="modal" data-target="#myModal" ng-click="testInterface(i._id)" ng-if="!i.testStatus" class="label label-warning" style="cursor:pointer;" title="操作：点击进行接口测试并保存，可修改状态~&#10;描述：未连接后端服务器进行测试，可能无法正常使用。&#10;测试者：//i.testUser//&#10;测试日期：//i.testTime//&#10;消耗时间： //i.costTime//">未测试</span>
-              <span data-toggle="modal" data-target="#myModal" ng-click="testInterface(i._id)" ng-if="1==i.testStatus" class="label label-success" style="cursor:pointer;" title="操作：点击进行接口测试并保存，可修改状态~&#10;描述：已通过后端服务器测试。&#10;测试者：//i.testUser//&#10;测试日期：//i.testTime//&#10;消耗时间：//i.costTime//">测试通过</span>
-              <span data-toggle="modal" data-target="#myModal" ng-click="testInterface(i._id)" ng-if="-1==i.testStatus" class="label label-danger" style="cursor:pointer;" title="操作：点击进行接口测试并保存，可修改状态~&#10;描述：接口返回值未通过规则校验。&#10;测试者：//i.testUser//&#10;测试日期：//i.testTime//&#10;消耗时间：//i.costTime//">测试失败</span>
+              <span data-toggle="modal" data-target="#myModal" ng-click="testInterface(i._id)" ng-if="1==i.testStatus" class="label label-success" style="cursor:pointer;" title="操作：点击进行接口测试并保存，可修改状态~&#10;描述：已通过后端服务器测试。&#10;测试者：//i.testUser//&#10;测试日期：//i.testTime//&#10;消耗时间：//i.costTime//">通过</span>
+              <span data-toggle="modal" data-target="#myModal" ng-click="testInterface(i._id)" ng-if="-1==i.testStatus" class="label label-danger" style="cursor:pointer;" title="操作：点击进行接口测试并保存，可修改状态~&#10;描述：接口返回值未通过规则校验。&#10;测试者：//i.testUser//&#10;测试日期：//i.testTime//&#10;消耗时间：//i.costTime//">失败</span>
             </td>
           </tr>
           {{#if editable}}
             <tr>
               <td colspan="10" class="text-center">
-                <a href="/schema/index.html?pid=//m.pid//&mid=//m._id//&projectName={{projectName}}&moduleName=//m.name//" target="_blank" class="glyphicon glyphicon-plus text-success">新增接口</a>
+                <a href="/interface/index.html?pid=//m.pid//&mid=//m._id//&projectName=//projectName//&moduleName=//m.name//" target="_blank" class="glyphicon glyphicon-plus text-success">新增接口</a>
               </td>
             </tr>
           {{/if}}
